fix: Update generate to use new Release feature for non-GitHub repos.
