+++ tests/generate_github/knope.toml
@@ -13,3 +13,7 @@ command = "git add . && git commit -m \"chore: prepare release $version\" && git
 
 [[workflows.steps]]
 type = "Release"
+
+[github]
+owner = "knope-dev"
+repo = "knope"
